refactor: update import path for fetchPokemonData and add new service file
